Split PlotWindow in PlotAction and PlotWindow
+ Add CopyAction
